+++ src/test/java/io/bootique/tapestry/testapp3/pages/Page1.java
@@ -1,0 +1,15 @@
+package io.bootique.tapestry.testapp3.pages;
+
+import io.bootique.tapestry.testapp3.services.DeferredService;
+import org.apache.tapestry5.ioc.annotations.Inject;
+
+public class Page1 {
+
+    @Inject
+    private DeferredService service;
+
+    public String getData() {
+        return service.data();
+    }
+
+}