+++ src/test/java/io/bootique/tapestry/testapp3/services/DeferredService.java
@@ -1,0 +1,6 @@
+package io.bootique.tapestry.testapp3.services;
+
+public interface DeferredService {
+
+    String data();
+}