+++ src/test/java/io/bootique/tapestry/testapp3/services/DeferredServiceImpl.java
@@ -1,0 +1,9 @@
+package io.bootique.tapestry.testapp3.services;
+
+public class DeferredServiceImpl implements DeferredService {
+
+    @Override
+    public String data() {
+        return ":DeferredServiceImpl:";
+    }
+}